+++ src/Policies/PendingFriendPolicy.php
@@ -49,8 +49,8 @@ class PendingFriendPolicy
     public function view($user, PendingFriend $pending): Response
     {
         return ($this->messenger->providerHasFriends()
-            && (string) $this->messenger->getProviderId() === (string) $pending->recipient_id
-            && $this->messenger->getProviderClass() === $pending->recipient_type)
+            && (string) $this->messenger->getProvider()->getKey() === (string) $pending->recipient_id
+            && $this->messenger->getProvider()->getMorphClass() === $pending->recipient_type)
             ? $this->allow()
             : $this->deny('Not authorized to view pending friend request');
     }
@@ -65,8 +65,8 @@ class PendingFriendPolicy
     public function update($user, PendingFriend $pending): Response
     {
         return ($this->messenger->providerHasFriends()
-            && (string) $this->messenger->getProviderId() === (string) $pending->recipient_id
-            && $this->messenger->getProviderClass() === $pending->recipient_type)
+            && (string) $this->messenger->getProvider()->getKey() === (string) $pending->recipient_id
+            && $this->messenger->getProvider()->getMorphClass() === $pending->recipient_type)
             ? $this->allow()
             : $this->deny('Not authorized to accept pending friend request');
     }
@@ -81,8 +81,8 @@ class PendingFriendPolicy
     public function delete($user, PendingFriend $pending): Response
     {
         return ($this->messenger->providerHasFriends()
-            && (string) $this->messenger->getProviderId() === (string) $pending->recipient_id
-            && $this->messenger->getProviderClass() === $pending->recipient_type)
+            && (string) $this->messenger->getProvider()->getKey() === (string) $pending->recipient_id
+            && $this->messenger->getProvider()->getMorphClass() === $pending->recipient_type)
             ? $this->allow()
             : $this->deny('Not authorized to deny pending friend request');
     }